==> 00 - Retrieve data.ipynb <==
import glob
import tqdm
import lxml.etree as ET
import csv
import collections
import os
from typing import List

# Parse data

models = ['data0',
         'data-bad3',
         'data-bad2',
         'data1',
         'data-bad1',
         'data2',
         'data-bad0',
         'data-CREMMA-Medieval',
         'data-manumffrench2']

COLS = ["model", "lang", "manuscript", "file", 
                                     "page_id", "line_id", 
                                     "baseline", 
                                     "transcription"]
Row = collections.namedtuple("Row", ["model", "lang", "manuscript", "file", 
                                     "page_id", "line_id", 
                                     "baseline", 
                                     "transcription"])

PageIds = collections.defaultdict(dict)
LineIds = collections.defaultdict(lambda: collections.defaultdict(dict))

def read_file(file: str, split_string: str = ".transc", model=None, lang=None) -> List[Row]:
    if not lang:
        lang, infered_model, manuscript, *_ = file.split(os.sep)
    else:
        manuscript = os.path.dirname(file).split(os.sep)[-1]
        
    if not model:
        model = infered_model
    data = []
    try:
        xml = ET.parse(file)
        fname = os.path.basename(file).split(split_string)[0]
        # Page ID
        if fname in PageIds[manuscript]:
            pid = PageIds[manuscript][fname]
        else:
            pid = PageIds[manuscript][fname] = len(PageIds[manuscript])
            
        for line in xml.findall("//{*}TextBlock/{*}TextLine"):
                
                
            text = " ".join([str(content.attrib.get("CONTENT", "")) or "" for content in line.findall(".//{*}String")])
            
            if not text:
                continue
            
            baseline = str(line.attrib["BASELINE"])
            # Line ID
            if baseline in LineIds[manuscript][fname]:
                lid = LineIds[manuscript][fname][baseline]
            else:
                lid = LineIds[manuscript][fname][baseline] = len(LineIds[manuscript][fname])
                
            data.append(Row(
                model,
                lang,
                manuscript,
                fname,
                pid,
                lid,
                baseline,
                text or ""))
    except Exception as E:
        print(E)
    return data

inputs = [
    file
    for file in glob.glob(f"*/*/*/*transc*.xml")
]
data = []
for file in tqdm.tqdm(inputs):
    data.extend(read_file(file))

## Retrieve GT

for file in tqdm.tqdm(glob.glob(f"../cremma/CREMMA-Medieval-LAT/data/*/*mufi*.xml")):
    data.extend(read_file(file, split_string=".mufi", model="GT", lang="lat"))
    
for file in tqdm.tqdm(glob.glob(f"../cremma-medieval/*/*mufi*.xml")):
    
    data.extend(read_file(file, split_string=".mufi", model="GT", lang="fro"))

# Export

import pandas as pd

df = pd.DataFrame(data, columns=COLS)
df.to_csv("annotations.csv.gzip", compression="gzip")

df.head()

df.tail()

df.manuscript.unique()
