==> 02 - Feature extraction.ipynb <==
import tqdm
import pandas as pd
from collections import Counter
from unicodedata import normalize

BIN_SIZE = 10

def get_ngrams(string: str, _sizes = (3, 4, 5), prefix="$"):
    ret = {}
    #string = normalize("NFC", string)
    for S in _sizes:
        nb = len(string) - (S-1)
        ret.update({
            prefix+k.replace(" ", "_"): v/nb
            for k, v in Counter([string[i:i+S] for i in range(len(string)-S+1)]).items()
        })
    return ret


df = pd.read_csv("cer.csv.gzip", compression="gzip", index_col=0)

print(df.shape)
df.drop_duplicates(subset=["manuscript", "page_id", "line_id", "transcription"], inplace=True)
print(df.shape)
df.head()

# Compute Bins

df["CER_BINS"] = ""
df.loc[df.CER < 10, "CER_BINS"] = "Good"
df.loc[df.CER.between(10, 25, inclusive="left"), "CER_BINS"] = "Acceptable"
df.loc[df.CER.between(25, 50, inclusive="left"), "CER_BINS"] = "Bad"
df.loc[df.CER >= 50, "CER_BINS"] = "Very bad"
df.groupby("CER_BINS").size()

df['CER_BINS'].unique()

# Compute ngrams

new_df = []

total = Counter()


for idx, row in tqdm.tqdm(df.iterrows(), total=len(df)):
    if len(str(row.transcription)) > 15:
        cnt = get_ngrams(str(row.transcription).lower())
        total += cnt
        new_df.append({
            "idx": idx,
            "lang": row.lang,
            "bin": row.CER_BINS,
            "manuscript": row.manuscript,
            "page_id": row.page_id,
            "line_id": row.line_id,
            "transcription": row.transcription, 
            "CER": row.CER,
            **cnt
        })
    
del df  # For memory sake

def ngrams(row):
    return get_ngrams(str(row[1].transcription).lower())
from multiprocessing import Pool

all_ngrams = []
with Pool(processes=4) as pool:
    # print same numbers in arbitrary order
    for results in tqdm.tqdm(pool.imap_unordered(ngrams, df.iterrows()), total=len(df)):
        all_ngrams.append(results)

## Cut ngrams appearing in less than .5% of the lines

total_presence = Counter([
    key
    for row in tqdm.tqdm(all_ngrams)
    for key in row 
    if key.startswith("$")
])
print(len(total_presence))
#filter_value = int(len(total_presence)/100)
#max_value = total_presence.most_common(filter_values)[-1][1]
#print(f"1% {filter_value}")

most_common = [key for key, _ in total_presence.most_common(2000)]
total_presence = {
    key: value
    for key, value in tqdm.tqdm(total_presence.items())
    if key in most_common
}
print(most_common)
#del total

# Sentences

import numpy as np
import random

random.shuffle(new_df)
KEPT = ["idx", "bin", "lang", "transcription", "manuscript", "page_id", "line_id", "CER", *total_presence.keys()]
# Optim
DTYPES = {
    "idx": np.int32,
    "page_id": np.int16,
    "line_id": np.int16,
    "CER": np.float16,
    **{
        feat: np.float16
        for feat in total_presence
    }
}

df = pd.DataFrame({
    key: pd.Series([
        row.get(key, np.nan)
        for row in new_df
    ], dtype=DTYPES.get(key))
    for key in KEPT})#, dtype={feature: np.float32 for feature in total_presence})
df.head()

df.dtypes

df.info(memory_usage=True)

df.to_hdf("features.hdf5", key='df')
df.head()

df["$abi"].value_counts()
df

df.bin.unique()

df["idx 	bin 	transcription 	manuscript 	page_id 	line_id 	CER".split()].to_hdf("texts.hdf5", key='df')
